Major change. I switched all size_t to unsigned int. This will allow me to directly pass vector contents to the GPU. The next change will be to have MPI operate directly on std::vectors so im not allocating extra memory and moving data around too many times for communication

+++ src/cl_kernels/boundary_conditions.cl
@@ -11,7 +11,7 @@ enforceBoundaryConditions(
                    float cur_time \n
 )  \n
 {   \n
-    size_t i = get_global_id(0);    \n
+    uint i = get_global_id(0);    \n
 \n
     if(i < nb_bnd) {    \n
         int j = bndry_indices[i];
